Issue templates for the Grml repository
With issue templates we can standardize the information we'd like
contributors to include when they open issues in our repository.

We want to help our contributors to provide specific, structured
information when they open issues.

This makes it easier for us to address and fix the issue.

Read more: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/about-issue-and-pull-request-templates